feat: Create personalize command
